+++ changelog.md
@@ -10,6 +10,7 @@
 ### Changes
 
 - [#3529](https://github.com/ignite/cli/pull/3529) Refactor plugin system to use gRPC
+- [#3750](https://github.com/ignite/cli/pull/3750) Update default Ignite network app to `v0.2.0`
 - [#3751](https://github.com/ignite/cli/pull/3751) Rename label to skip changelog check
 - [#3745](https://github.com/ignite/cli/pull/3745) Set tx fee amount as option
 - [#3748](https://github.com/ignite/cli/pull/3748) Change default rpc endpoint to a working one
@@ -23,6 +24,7 @@
 - [#3536](https://github.com/ignite/cli/pull/3536) Change app.go to v2 and add AppWiring feature
 - [#3659](https://github.com/ignite/cli/pull/3659) cosmos-sdk `v0.50.x`
 - [#3670](https://github.com/ignite/cli/pull/3670) Remove nodetime binaries
+- [#3724](https://github.com/ignite/cli/pull/3724) Add or vendor proto packages from Go dependencies
 - [#3715](https://github.com/ignite/cli/pull/3715) Add test suite for the cli tests
 
 ### Changes